move JobManagerConfig to configuration package
