+++ VERSION
@@ -1,1 +1,1 @@
-1.19.0-A1
+1.18.4
